bugfix@BC-1500: Audit Issue#7 Prefer explicit unsigned integer sizes

+++ contracts/compliance/ComplianceConfigurationService.sol
@@ -36,7 +36,7 @@ contract ComplianceConfigurationService is IDSComplianceConfigurationService, Co
     function setCountriesCompliance(string[] calldata _countries, uint256[] calldata _values) public override onlyTransferAgentOrAbove {
         require(_countries.length <= 35, "Exceeded the maximum number of countries");
         require(_countries.length == _values.length, "Wrong length of parameters");
-        for (uint i = 0; i < _countries.length; i++) {
+        for (uint8 i = 0; i < _countries.length; i++) {
             setCountryCompliance(_countries[i], _values[i]);
         }
     }
